# swagger.yaml
openapi: 3.0.0
info:
  title: API Quản lý Nông Trại Nấm
  version: 1.0.0
  description: Tài liệu API cho Hệ thống Quản lý Nuôi trồng Nấm
servers:
  - url: /
    description: Máy chủ đang phục vụ Swagger (tự dùng host và cổng hiện tại)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Pagination:
      type: "object"
      properties:
        totalItems:
          type: "integer"
        currentPage:
          type: "integer"
          minimum: 1
        totalPages:
          type: "integer"
        pageSize:
          type: "integer"
          minimum: 1
    TaskSubmissionSummary:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        status:
          type: "string"
          enum: ["PENDING","APPROVED","REJECTED","CANCELLED"]
        submittedByUserId:
          type: "integer"
          nullable: true
          readOnly: true
        reviewedByUserId:
          type: "integer"
          nullable: true
          readOnly: true
        submittedAt:
          type: "string"
          format: "date-time"
        reviewedAt:
          type: "string"
          format: "date-time"
          nullable: true
        reason:
          type: "string"
          nullable: true
    TaskSubmission:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        status:
          type: "string"
          enum: ["PENDING","APPROVED","REJECTED","CANCELLED"]
        submittedByUserId:
          type: "integer"
          nullable: true
          readOnly: true
        reviewedByUserId:
          type: "integer"
          nullable: true
          readOnly: true
        submittedAt:
          type: "string"
          format: "date-time"
        reviewedAt:
          type: "string"
          format: "date-time"
          nullable: true
        reason:
          type: "string"
          nullable: true
        taskId:
          type: "string"
          format: "uuid"
        notes:
          type: "string"
          nullable: true
        evidence:
          type: "array"
          items:
            type: "object"
            properties:
              type:
                type: "string"
                enum: ["CARE_LOG","GROWTH_PROGRESS","HARVEST"]
              recordId:
                type: "integer"
                minimum: 1
              record:
                oneOf: [{"$ref":"#/components/schemas/CultivationCareLog"},{"$ref":"#/components/schemas/GrowthProgressRecord"},{"$ref":"#/components/schemas/HarvestRecord"}]
    Task:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        title:
          type: "string"
        description:
          type: "string"
          nullable: true
        status:
          type: "string"
          enum: ["TODO","IN_PROGRESS","PENDING_REVIEW","COMPLETED","CANCELLED"]
        batchId:
          type: "integer"
          nullable: true
        batchCode:
          type: "string"
          nullable: true
        assigneeUserId:
          type: "integer"
          nullable: true
          readOnly: true
        assignee:
          type: "object"
          nullable: true
          properties:
            id:
              type: "integer"
              minimum: 1
            username:
              type: "string"
            full_name:
              type: "string"
            role:
              type: "string"
              enum: ["admin","manager","staff"]
        createdByUserId:
          type: "integer"
          nullable: true
          readOnly: true
        dueAt:
          type: "string"
          format: "date-time"
          nullable: true
        createdAt:
          type: "string"
          format: "date-time"
        updatedAt:
          type: "string"
          format: "date-time"
        version:
          type: "integer"
          readOnly: true
        submissions:
          type: "array"
          items:
            $ref: "#/components/schemas/TaskSubmissionSummary"
          description: "List/create/patch: tối đa một lần gửi gần nhất, không có evidence/notes."
    TaskDetail:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        title:
          type: "string"
        description:
          type: "string"
          nullable: true
        status:
          type: "string"
          enum: ["TODO","IN_PROGRESS","PENDING_REVIEW","COMPLETED","CANCELLED"]
        batchId:
          type: "integer"
          nullable: true
        batchCode:
          type: "string"
          nullable: true
        assigneeUserId:
          type: "integer"
          nullable: true
          readOnly: true
        assignee:
          type: "object"
          nullable: true
          properties:
            id:
              type: "integer"
              minimum: 1
            username:
              type: "string"
            full_name:
              type: "string"
            role:
              type: "string"
              enum: ["admin","manager","staff"]
        createdByUserId:
          type: "integer"
          nullable: true
          readOnly: true
        dueAt:
          type: "string"
          format: "date-time"
          nullable: true
        createdAt:
          type: "string"
          format: "date-time"
        updatedAt:
          type: "string"
          format: "date-time"
        version:
          type: "integer"
          readOnly: true
        submissions:
          type: "array"
          items:
            $ref: "#/components/schemas/TaskSubmission"
          description: "Toàn bộ lịch sử với snapshot minh chứng."
...
    TaskCreateInput:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
          nullable: true
        batchId:
          type: integer
          nullable: true
        assigneeUserId:
          type: integer
          nullable: true
        dueAt:
          type: string
          format: date-time
          nullable: true
      required: [title]
    TaskUpdateInput:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
          nullable: true
        batchId:
          type: integer
          nullable: true
        assigneeUserId:
          type: integer
          nullable: true
        dueAt:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [TODO, IN_PROGRESS, CANCELLED]
...
    CultivationCareLog:
      type: object
      properties:
        id:
          type: integer
        batchId:
          type: integer
        actionType:
          type: string
        notes:
          type: string
        recordedAt:
          type: string
          format: date-time
        createdByUserId:
          type: "integer"
          nullable: true
          readOnly: true
        updatedByUserId:
          type: "integer"
          nullable: true
          readOnly: true
    GrowthProgressRecord:
      type: object
      required: [id, batchId, stage, notes, recordedAt, createdAt, updatedAt]
      properties:
        id:
          type: integer
          readOnly: true
        batchId:
          type: integer
          readOnly: true
        stage:
          type: string
        notes:
          type: string
        recordedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
        images:
          type: array
          items:
            $ref: '#/components/schemas/GrowthProgressImage'
        createdByUserId:
          type: "integer"
          nullable: true
          readOnly: true
        updatedByUserId:
          type: "integer"
          nullable: true
          readOnly: true
    GrowthProgressImage:
      type: object
      required: [id, growthProgressId, imageUrl, originalName, mimeType, fileSize, createdAt]
      properties:
        id:
          type: integer
          readOnly: true
        growthProgressId:
          type: integer
          readOnly: true
        imageUrl:
          type: string
          description: Duong dan cong khai cua anh thuc te
        originalName:
          type: string
        mimeType:
          type: string
          example: image/jpeg
        fileSize:
          type: integer
          description: Kich thuoc tep tinh bang byte
        createdAt:
          type: string
          format: date-time
          readOnly: true
...
    HarvestRecord:
      type: object
      properties:
        id:
          type: integer
        batchId:
          type: integer
        harvestedAt:
          type: string
          format: date-time
        totalYieldKg:
          type: number
        qualityGrade:
          type: string
        notes:
          type: string
        createdByUserId:
          type: "integer"
          nullable: true
          readOnly: true
        updatedByUserId:
          type: "integer"
          nullable: true
          readOnly: true
...
security:
  - bearerAuth: []
paths:
...
  /api/dashboard/tasks:
    get:
      summary: "Danh sách công việc; mặc định TODO, IN_PROGRESS, PENDING_REVIEW"
      tags: ["Dashboard"]
      parameters: [{"name":"page","in":"query","schema":{"type":"integer","minimum":1,"default":1}},{"name":"limit","in":"query","schema":{"type":"integer","minimum":1,"default":10,"maximum":50}},{"name":"search","in":"query","schema":{"type":"string"}},{"name":"batchId","in":"query","schema":{"type":"integer","minimum":1}},{"name":"assigneeUserId","in":"query","schema":{"type":"integer","minimum":1}},{"name":"status","in":"query","schema":{"type":"string","enum":["TODO","IN_PROGRESS","PENDING_REVIEW","COMPLETED","CANCELLED","ALL"]}}]
      responses:
        "200":
          description: "Thành công"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  data:
                    type: "array"
                    items:
                      $ref: "#/components/schemas/Task"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "400":
          description: "Dữ liệu không hợp lệ"
        "401":
          description: "Token thiếu hoặc không hợp lệ"
        "403":
          description: "Không đủ quyền"
        "404":
          description: "Không tìm thấy dữ liệu"
        "409":
          description: "Dữ liệu/trạng thái thay đổi; tải lại trước khi thử lại"
        "500":
          description: "Lỗi máy chủ"
    post:
      summary: "Tạo việc TODO (manager/admin)"
      tags: ["Dashboard"]
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreateInput"
      responses:
        "201":
          description: "Thành công"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  message:
                    type: "string"
                  data:
                    $ref: "#/components/schemas/Task"
        "400":
          description: "Dữ liệu không hợp lệ"
        "401":
          description: "Token thiếu hoặc không hợp lệ"
        "403":
          description: "Không đủ quyền"
        "404":
          description: "Không tìm thấy dữ liệu"
        "409":
          description: "Dữ liệu/trạng thái thay đổi; tải lại trước khi thử lại"
        "500":
          description: "Lỗi máy chủ"
  /api/dashboard/tasks/{id}:
    get:
      summary: "Chi tiết công việc và lịch sử minh chứng"
      tags: ["Dashboard"]
      parameters: [{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}]
      responses:
        "200":
          description: "Thành công"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  message:
                    type: "string"
                  data:
                    $ref: "#/components/schemas/TaskDetail"
        "400":
          description: "Dữ liệu không hợp lệ"
        "401":
          description: "Token thiếu hoặc không hợp lệ"
        "403":
          description: "Không đủ quyền"
        "404":
          description: "Không tìm thấy dữ liệu"
        "409":
          description: "Dữ liệu/trạng thái thay đổi; tải lại trước khi thử lại"
        "500":
          description: "Lỗi máy chủ"
    patch:
      summary: "Cập nhật việc; không hoàn thành/chờ duyệt trực tiếp"
      tags: ["Dashboard"]
      parameters: [{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdateInput"
      responses:
        "200":
          description: "Thành công"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  message:
                    type: "string"
                  data:
                    $ref: "#/components/schemas/Task"
        "400":
          description: "Dữ liệu không hợp lệ"
        "401":
          description: "Token thiếu hoặc không hợp lệ"
        "403":
          description: "Không đủ quyền"
        "404":
          description: "Không tìm thấy dữ liệu"
        "409":
          description: "Dữ liệu/trạng thái thay đổi; tải lại trước khi thử lại"
        "500":
          description: "Lỗi máy chủ"
    delete:
      summary: "Xóa việc chưa có lịch sử minh chứng (manager/admin)"
      tags: ["Dashboard"]
      parameters: [{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}]
      responses:
        "200":
          description: "Đã xóa"
        "400":
          description: "Dữ liệu không hợp lệ"
        "401":
          description: "Token thiếu hoặc không hợp lệ"
        "403":
          description: "Không đủ quyền"
        "404":
          description: "Không tìm thấy dữ liệu"
        "409":
          description: "Dữ liệu/trạng thái thay đổi; tải lại trước khi thử lại"
        "500":
          description: "Lỗi máy chủ"
...